для наставника
