+++ public/index.html
@@ -47,6 +47,6 @@
           </section>
       </div>
     </main>
-    <script src="../src/main.js"></script>
+    <script src="../src/bundle.js"></script>
   </body>
 </html>
